Add web-safe AI service handling and UI messaging
